+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 * Implementation of tokens management
 * Serialization - int management
 * Coingecko: Get historical market data include price, market cap, and 24h volume within a range of timestamp
+* Check curve type in addressFormatControl method
 
 #### Version 1.0.23 (2022-07-21)
 * BREAKING-CHANGE - Implement VarInt when getting signature payload